Adds packages, tweaks router, adds Redux boilerplate
We will be able to do more with Redux when we have more components.
But we should now be ready to think about the Redux store when we create
components.

Components should get their state from Redux, usually, instead of from
parent components.

As a result components can be functional and stateless if we want. :-)

[closes #34]